call monitor before allocating region on mmap
the idea of the region_mapping is, that the monitor can hint elkvm on
how to proceed with the mmap. with the old order - allocate memory, call
monitor - it was not possible to resize the mappings.

additionally with the page table options field inside the region_mapping
it is now possible for the monitor to map pages read-only, or without
execution rights etc.

semantically, if the guest_virt field in the region_mapping is set to
0x0, ELKVM will decide on a suitable guest address itself. the host
address is always overwritten.