create a resized mapping if there is not enough room
for calls to mremap(), when new_size is larger than we can fit in the
existing mapping we now create a new mapping, copy everything over and
return the guest address of the new mapping.

while this may be slower than just appending the mapping it makes it
possible to use mremap() on the same mapping multiple times.