Move defintion of elkvm_syscalls table
Defining a variable as 'static' in a header file will create
static instances of this variable in every object that includes this
header! We really want only one instance of this table and it is
actually only used in a single object, so move the definition into this
object file.