Mem leak fix: create_vm()
Deal with a memory leak / double free that I saw from creating ELKVMs.
Previously we created VMs by emplace'ing them into Elkvm::vmi and then
creating a shared_ptr() to a pointer to an element in this list. This
shared_ptr was improperly freed twice in some tests I did.

My fix changes the list of ELKVMs to a list of shared ptrs of ELKVMs.
Therefore, the VMs are released only when the list is freed (aka at
termination time) and only released once. Removing the emplace_back()
should not be critical as ELKVM is not allocating thousands of such
objects all the time.